Moved CodecConfiguration from ChannelBindingAdapterConfiguration -> spring-cloud-streams-codec
