+++ spring-cloud-streams-codec/src/main/java/org/springframework/cloud/streams/config/CodecConfiguration.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2015 the original author or authors.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.streams.config;
+
+import java.util.ArrayList;
+import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.xd.dirt.integration.bus.serializer.MultiTypeCodec;
+import org.springframework.xd.dirt.integration.bus.serializer.kryo.FileKryoRegistrar;
+import org.springframework.xd.dirt.integration.bus.serializer.kryo.KryoRegistrar;
+import org.springframework.xd.dirt.integration.bus.serializer.kryo.PojoCodec;
+
+/**
+ * @author David Turanski
+ */
+@Configuration
+public class CodecConfiguration {
+
+
+	@Autowired
+	ApplicationContext applicationContext;
+
+	@ConditionalOnMissingBean(KryoCodecProperties.class)
+	@Bean(name = "spring.cloud.streams.codec.kryo.CONFIGURATION_PROPERTIES")
+	public KryoCodecProperties kryoCodecProperties() {
+		return new KryoCodecProperties();
+	}
+
+	@Bean
+	@ConditionalOnMissingBean(name = "codec")
+	public MultiTypeCodec<?> codec() {
+		Map<String, KryoRegistrar> kryoRegistrarMap = applicationContext.getBeansOfType(KryoRegistrar
+				.class);
+		return new PojoCodec(new ArrayList<>(kryoRegistrarMap.values()), kryoCodecProperties().isReferences());
+	}
+
+	@Bean
+	public KryoRegistrar fileRegistrar() {
+		return new FileKryoRegistrar();
+	}
+}
