+++ spring-cloud-streams-codec/src/main/java/org/springframework/cloud/streams/config/KryoCodecProperties.java
